Modularizza missioni e scheduler
