+++ bot_app/__init__.py
@@ -1,0 +1,8 @@
+"""Utility per inizializzare il bot suddivise per moduli."""
+
+from .bootstrap import BotAppContext, create_app_context
+
+__all__ = [
+    "BotAppContext",
+    "create_app_context",
+]
